+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to `dropbox-api` will be documented in this file
 
+## 1.5.0 - 2017-07-17
+
+- add `revokeToken` and `getAccountInfo`
+
+## 1.4.0 - 2017-07-11
+
+- add `listSharedLinks`
+
+## 1.3.0 - 2017-07-04
+
+- add error code to thrown exception
+
 ## 1.2.0 - 2017-04-29
 
 - added `createSharedLinkWithSettings`